Z.3-CHAMBER/GLYPH.BIT/_ARCHIVE/gyphbit-noctua/requirements.txt
python-telegram-bot>=21.0
openai>=1.30.0
python-dotenv==1.0.0

